Introduce support for audio from NDI sources
Sources can now have their audio output through the default (or user
selected) playback device. Note that all sources are created to be muted
by default.

Volume can be changed on a per-source basis, and is mixed with all other
sources -- except if the "Audio" > "Focused Window Only" option is
enabled, in which then only the focused source window's audio will be
played.